+++ src/include/sll/api/_generated.h
@@ -814,6 +814,102 @@
 
 
 
+#define __SLL_API_TYPE_sll_api_string_index __SLL_CHECK_OUTPUT sll_integer_t
+#define __SLL_API_ARGS_sll_api_string_index const sll_string_t* a,const sll_object_t* b
+/**
+ * \flags check_output func optimizable
+ * \name sll_api_string_index
+ * \group string-api
+ * \desc Docs!
+ * \arg const sll_string_t* a -> 
+ * \arg const sll_object_t* b -> 
+ * \ret sll_integer_t
+ */
+/**
+ * \flags check_output func optimizable
+ * \name sll_api_string_index_raw
+ * \group raw-api
+ * \subgroup raw-api-string-api
+ * \desc Wrapper function for :sll_api_string_index:
+ * \arg const sll_object_t*const* al -> Arguments
+ * \arg sll_arg_count_t all -> Argument count
+ * \ret sll_object_t* -> The return value of the function
+ */
+
+
+
+#define __SLL_API_TYPE_sll_api_string_index_list __SLL_CHECK_OUTPUT sll_integer_t
+#define __SLL_API_ARGS_sll_api_string_index_list const sll_string_t* a,const sll_object_t* b
+/**
+ * \flags check_output func optimizable
+ * \name sll_api_string_index_list
+ * \group string-api
+ * \desc Docs!
+ * \arg const sll_string_t* a -> 
+ * \arg const sll_object_t* b -> 
+ * \ret sll_integer_t
+ */
+/**
+ * \flags check_output func optimizable
+ * \name sll_api_string_index_list_raw
+ * \group raw-api
+ * \subgroup raw-api-string-api
+ * \desc Wrapper function for :sll_api_string_index_list:
+ * \arg const sll_object_t*const* al -> Arguments
+ * \arg sll_arg_count_t all -> Argument count
+ * \ret sll_object_t* -> The return value of the function
+ */
+
+
+
+#define __SLL_API_TYPE_sll_api_string_index_reverse __SLL_CHECK_OUTPUT sll_integer_t
+#define __SLL_API_ARGS_sll_api_string_index_reverse const sll_string_t* a,const sll_object_t* b
+/**
+ * \flags check_output func optimizable
+ * \name sll_api_string_index_reverse
+ * \group string-api
+ * \desc Docs!
+ * \arg const sll_string_t* a -> 
+ * \arg const sll_object_t* b -> 
+ * \ret sll_integer_t
+ */
+/**
+ * \flags check_output func optimizable
+ * \name sll_api_string_index_reverse_raw
+ * \group raw-api
+ * \subgroup raw-api-string-api
+ * \desc Wrapper function for :sll_api_string_index_reverse:
+ * \arg const sll_object_t*const* al -> Arguments
+ * \arg sll_arg_count_t all -> Argument count
+ * \ret sll_object_t* -> The return value of the function
+ */
+
+
+
+#define __SLL_API_TYPE_sll_api_string_index_reverse_list __SLL_CHECK_OUTPUT sll_integer_t
+#define __SLL_API_ARGS_sll_api_string_index_reverse_list const sll_string_t* a,const sll_object_t* b
+/**
+ * \flags check_output func optimizable
+ * \name sll_api_string_index_reverse_list
+ * \group string-api
+ * \desc Docs!
+ * \arg const sll_string_t* a -> 
+ * \arg const sll_object_t* b -> 
+ * \ret sll_integer_t
+ */
+/**
+ * \flags check_output func optimizable
+ * \name sll_api_string_index_reverse_list_raw
+ * \group raw-api
+ * \subgroup raw-api-string-api
+ * \desc Wrapper function for :sll_api_string_index_reverse_list:
+ * \arg const sll_object_t*const* al -> Arguments
+ * \arg sll_arg_count_t all -> Argument count
+ * \ret sll_object_t* -> The return value of the function
+ */
+
+
+
 #define __SLL_API_TYPE_sll_api_string_pad_left void
 #define __SLL_API_ARGS_sll_api_string_pad_left const sll_string_t* a,const sll_integer_t b,const sll_read_char_t c,sll_string_t* out
 /**