+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 - Implemented `sll_ext:debug_get_name` ([#130] and [#137])
 - Implemented `sll_ext:debug_get_vm_config` ([#130] and [#132])
 - Implemented `sll_get_location` and `sll_ext:debug_get_location` ([#141])
+- Implemented `sll_string_index_reverse`, `sll_string_index_reverse_char`, `sll_string_index_reverse_multiple`, `sll_api_string_index_reverse` and `sll:string_index_reverse` ([#143])
+- Implemented `sll_string_index`, `sll_string_index_char`, `sll_string_index_multiple`, `sll_api_string_index` and `sll:string_index` ([#142])
 - Implemented `sys$library_path` ([#145])
 - New node type representing source code lines: `SLL_NODE_TYPE_DBG` ([#140])
 - VM call stack is now part of the public API
@@ -1209,6 +1211,8 @@ Unfortunately, no versions were assigned before 2021-6-15 (:disappointed:), so t
 [#146]: https://github.com/sl-lang/sll/issues/146
 [#145]: https://github.com/sl-lang/sll/issues/145
 [#144]: https://github.com/sl-lang/sll/issues/144
+[#143]: https://github.com/sl-lang/sll/issues/143
+[#142]: https://github.com/sl-lang/sll/issues/142
 [#141]: https://github.com/sl-lang/sll/issues/141
 [#140]: https://github.com/sl-lang/sll/issues/140
 [#139]: https://github.com/sl-lang/sll/issues/139